modifications

+++ 2023/Net8WhatsNew/code/WhatsNewEFCore8/WhatsNewEFCore8/Book.cs
@@ -1,4 +1,4 @@
 ﻿namespace WhatsNewEFCore8;
-public record Book(int ID ,string Name)
+public record Book(long ID ,string Name)
 {
 }
